Download PartyPlnr_Chatbot_GitHubReady/requirements.txt
flask
openai
python-dotenv
pandas
